+++ src/lib/lforms/error-messages.js
@@ -19,6 +19,9 @@ export const ErrorMessages = {
     'invalidTagInHelpHTMLContent': {
       'en': 'Invalid HTML tags/attributes found in the help text.'
     },
+    'invalidTagInLegalHTMLContent': {
+      'en': 'Invalid HTML tags/attributes found in the legal text.'
+    },
     'invalidTagInHTMLContent': {
       'en': 'Invalid HTML tags/attributes found in item text/prefix.'
     }